Fix CI: size_t portability and glibc compat shims
- Use the size_t platform typealias in native write/header callbacks and
  int64_t for postfields length, so 32-bit Android targets (where size_t is
  UInt) compile
- Add Linux glibc shims for fcntl64 and __explicit_bzero_chk/explicit_bzero
  (referenced by libcurl built against glibc >= 2.28/2.38, missing from the
  Kotlin/Native bundled sysroot) and guard the whole file with __linux__

+++ curl-kmp/native/curl_kmp_helpers.c
@@ -54,7 +54,7 @@ CURLcode curl_kmp_easy_setopt_headerfunction(CURL *easy, curl_write_callback cb,
  * the transfer runs. The native side keeps the ByteArray pinned; the JNI
  * bridge keeps an owned copy per handle. A NULL data resets the method back
  * to GET explicitly (curl keeps the POST flag otherwise). */
-CURLcode curl_kmp_easy_setopt_postfields(CURL *easy, const void *data, size_t len) {
+CURLcode curl_kmp_easy_setopt_postfields(CURL *easy, const void *data, int64_t len) {
     if (!data) {
         CURLcode rc = curl_easy_setopt(easy, CURLOPT_HTTPGET, 1L);
         if (rc != CURLE_OK) {
